Solved day 18, part 1

+++ src/main/java/nl/mvdr/adventofcode/adventofcode2017/day18/State.java
@@ -53,7 +53,7 @@ class State {
             throw new IllegalArgumentException(register);
         }
         
-        return registers.getOrDefault(register, Long.valueOf(0L)).intValue();
+        return registers.getOrDefault(register, Long.valueOf(0L)).longValue();
     }
     
     OptionalLong getRecoveredFrequency() {
